+++ backend/models/Category.js
@@ -7,6 +7,11 @@ const Category = sequelize.define('Category', {
     primaryKey:    true,
     autoIncrement: true,
   },
+  slug: {
+    type:      DataTypes.STRING,
+    allowNull: false,
+    unique:    true,
+  },
   name: {
     type:      DataTypes.STRING,
     allowNull: false,