+++ backend/server.js
@@ -2,6 +2,7 @@ import 'dotenv/config';
 import express, { json } from 'express';
 import sequelize from './config/database.js';
 import categorias from './routes/categorias.js';
+import produtos from './routes/produtos.js';
 import cors from 'cors';
 
 const app = express();
@@ -9,10 +10,11 @@ const app = express();
 app.use(json());
 
 app.use(cors({
-  origin: 'http://localhost:5173'
+  origin: 'http://localhost:5173',
 }));
 
 app.use('/api/categorias', categorias);
+app.use('/api/produtos', produtos);
 
 const PORT = process.env.PORT || 3000;